Travel_Planner/requirements.txt
rich
requests
google-search-results
python-dotenv
Flask